Issue #76: it IS learning (with the tiny image); it just takes up to 120 epochs before it makes progress.
Perhaps this is the phenomenon I wrote up about some time ago:
My experiments have seen cases where the loss function decreases steadily (albeit slowly) while the accuracy is stuck at zero, but then does a big jump to say 50 or more. I think this is because the continuous floats are slowly moving up to the cutoff point of 0.5 where a classification switches from being 0 to 1, which makes an 'all zeros' become 'approximately half are 1s', which could explain this behaviour.